Pull asThingType, asEntity etc. up into Concept
